Fix aiming issue
Up to this point, there was an annoying issue which made the tank barrel get stuck and jitter when moving it in certain directions. This was because the deprojection of the crosshair screen coordinates to world coordinates did not ignore the tank, leading the final world position to be on top of the tank. Once the tank was added to the list of ignored actors, the issue disappeared.

- Fix mortar firing. The projectile spawn location was too close to the barrel so the projectile exploded before launching.